requirements.txt
alembic==1.4.2
apispec==1.3.3
appdirs==1.4.4
argcomplete==1.11.1
arrow==0.15.7
asn1crypto==1.3.0
attrs==19.3.0
auditwheel==3.1.1
azure-common==1.1.25
azure-core==1.7.0
Babel==2.8.0
boto==2.49.0
boto3==1.13.23
botocore==1.16.23
cached-property==1.5.1
cachetools==4.1.1
cattrs==1.0.0
certifi==2020.4.5.1
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
chispa==0.6.0
click==7.1.2
cloudsync==1.5.3
cloudsync-onedrive==1.0.3
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.34
cx-Oracle==8.0.1
dataclasses==0.6
defusedxml==0.6.0
dill==0.3.2
distlib==0.3.1
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
dynaconf==2.2.3
email-validator==1.1.1
filelock==3.0.12
findspark==1.4.2
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-AutoIndex==0.6.6
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-Dropzone==1.5.4
Flask-Excel==0.0.7
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-Silk==0.2
Flask-SQLAlchemy==2.4.3
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
graphviz==0.14
gunicorn==19.10.0
hdfs==2.5.8
identify==1.4.25
idna==2.9
importlib-metadata==1.6.1
iso8601==0.1.12
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
json-merge-patch==0.2
jsonschema==3.2.0
kafka-python==2.0.1
lazy-object-proxy==1.5.0
lml==0.1.0
lockfile==0.12.2
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
more-itertools==8.4.0
msgpack==1.0.0
msrest==0.6.17
mysql-connector-python==8.0.21
natsort==7.0.1
nodeenv==1.4.0
numpy==1.18.5
oauthlib==3.1.0
onedrivesdk-fork==1.1.14
oscrypto==1.2.0
packaging==20.4
pandas==1.0.4
parquet==1.3.1
pathtools==0.1.2
pendulum==1.4.4
pluggy==0.13.1
ply==3.11
pre-commit==2.6.0
prison==0.1.3
protobuf==3.12.2
psutil==5.7.0
psycopg2-binary==2.8.5
py==1.10.0
py4j==0.10.7
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.9.8
pyelftools==0.26
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
Pygments==2.7.4
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
pystrict==1.1
pytest==5.4.3
pymssql==2.1.5
pytest-spark==0.6.0
python-box==3.4.6
python-daemon==2.2.4
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-gitlab==2.2.0
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
pytzdata==2019.3
PyYAML==5.4
quickxorhash==1.0.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.7
s3transfer==0.3.3
setproctitle==1.1.10
sh==1.13.1
six==1.15.0
smart-open==2.1.0
snowflake-connector-python==2.2.9
SQLAlchemy==1.3.18
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.8
tabulate==0.8.7
tenacity==4.12.0
text-unidecode==1.3
texttable==1.6.3
thrift==0.13.0
thriftpy2==0.4.11
toml==0.10.1
typing-extensions==3.7.4.2
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.9
virtualenv==20.0.28
watchdog==0.10.3
wcwidth==0.2.4
Werkzeug==1.0.1
WTForms==2.3.3
xlrd==1.2.0
xxhash==1.4.4
zipp==3.1.0
zope.deprecation==4.4.0
PyHive==0.6.4